requirements.txt
asgiref==3.7.2
Django==4.1.5
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
pytz==2023.4
PyYAML==6.0
sqlparse==0.4.4
tzdata==2023.4
requests==2.28.2